// src/app/modules/products/enums/product.enum.ts
export enum ProductActionTypes {
  FETCH_PRODUCT = '[PRODUCTS] FETCH_PRODUCT',
  FETCH_PRODUCT_SUCCESS = '[PRODUCTS] FETCH_PRODUCT_SUCCESS',
  FETCH_PRODUCT_FAILURE = '[PRODUCTS] FETCH_PRODUCT_FAILURE',
  CREATE_PRODUCT = '[PRODUCTS] CREATE_PRODUCT',
  CREATE_PRODUCT_SUCCESS = '[PRODUCTS] CREATE_PRODUCT_SUCCESS',
  CREATE_PRODUCT_FAILURE = '[PRODUCTS] CREATE_PRODUCT_FAILURE',
  UPDATE_PRODUCT = '[PRODUCTS] UPDATE_PRODUCT',
  UPDATE_PRODUCT_SUCCESS = '[PRODUCTS] UPDATE_PRODUCT_SUCCESS',
  UPDATE_PRODUCT_FAILURE = '[PRODUCTS] UPDATE_PRODUCT_FAILURE',
  DELETE_PRODUCT = '[PRODUCTS] DELETE_PRODUCT',
  DELETE_PRODUCT_SUCCESS = '[PRODUCTS] DELETE_PRODUCT_SUCCESS',
  DELETE_PRODUCT_FAILURE = '[PRODUCTS] DELETE_PRODUCT_FAILURE',
}